move to better suite; add solidity sources
