+++ test-clients/src/main/resource/contract/solidity/SubLevelTransferringContract.sol
@@ -1,0 +1,14 @@
+pragma solidity ^0.8.0;
+
+contract SubLevel {
+
+    constructor() payable {}
+
+    function receiveTinybars() public payable returns (bool) {
+        return true;
+    }
+
+    function nonPayableReceive() public pure returns (bool) {
+        return true;
+    }
+}
